🔥 remove: CalendarSection 컴포넌트 파일로 분리
